DESARROLLO-A3/1-agente-conversacional/requirements.txt
flask==3.1.2
httpx==0.28.1
python-dotenv==1.2.2
openpyxl==3.1.5
pypdf==6.9.1
